Add files via upload
In my final year NLP project, I designed an innovative chatbot tailored for our educational institution. This chatbot offers a dynamic user experience by accepting both voice and text inputs, adapting to the preferences of the user. To ensure the efficiency and accuracy of the responses, the system undergoes several essential preprocessing steps. These include lemmatization, tokenization, and vectorization, which collectively enhance the understanding and context of the input queries.

The core of this chatbot lies in its machine learning model, carefully trained on a diverse dataset to provide the most relevant and contextually appropriate responses. Through natural language processing techniques, it evaluates the user's query and retrieves the best-matching response from a predefined set of answers.

To make this innovative tool accessible to users, I seamlessly integrated it into a web application using Flask, a popular Python web framework. This integration not only provides a user-friendly interface for interacting with the chatbot but also allows users to access it from any device with internet connectivity. The web-based platform enhances user engagement and convenience, making it a valuable addition to our institute's digital resources.

This project represents a culmination of natural language processing, machine learning, and web development, bringing together a responsive and intelligent chatbot that addresses user inquiries and contributes to a more interactive and informative user experience within our educational institution.